requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
Authlib==0.13
cffi==1.13.2
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.8
enum34==1.1.6
funcsigs==1.0.2
futures==3.1.1
grpcio==1.26.0
grpcio-tools==1.26.0
grpcio-status==1.26.0
importlib-metadata==0.23
ipaddress==1.0.23
more-itertools==5.0.0
msgpack-python==0.5.6
packaging==19.2
pathlib2==2.3.5
pluggy==0.13.0
protobuf==3.11.1
py==1.8.0
pycparser==2.19
pyparsing==2.4.3
pytest==4.6.6
requests==2.22.0
scandir==1.10.0
six==1.13.0
wcwidth==0.1.7
zipp==0.6.0
# vmdgrpc==0.2.0
# vmdauthutils==0.1.0
OSlash==0.5.1